feat: Don't run validation if the video already validated.

+++ posthog/temporal/ai/session_summary/activities/videos.py
@@ -1,5 +1,6 @@
 import json
-from typing import cast
+from dataclasses import asdict
+from typing import Any, cast
 
 import temporalio
 from temporalio.exceptions import ApplicationError
@@ -8,6 +9,8 @@ from posthog.models.user import User
 from posthog.sync import database_sync_to_async
 from posthog.temporal.ai.session_summary.types.single import SingleSessionSummaryInputs
 
+from ee.hogai.session_summaries.session.output_data import SessionSummarySerializer
+from ee.hogai.session_summaries.session.video_validate_session import SessionSummaryVideoValidator
 from ee.models.session_summaries import SingleSessionSummary
 
 
@@ -15,7 +18,7 @@ from ee.models.session_summaries import SingleSessionSummary
 async def validate_llm_single_session_summary_with_videos_activity(
     inputs: SingleSessionSummaryInputs,
 ) -> None:
-    """Validate the LLM single session summary with videos"""
+    """Validate the LLM single session summary with videos"""  # TODO: Don't do video check if it's already done (get from run metadata)
     summary_row = await database_sync_to_async(SingleSessionSummary.objects.get_summary, thread_sensitive=False)(
         team_id=inputs.team_id,
         session_id=inputs.session_id,
@@ -26,16 +29,39 @@ async def validate_llm_single_session_summary_with_videos_activity(
             f"Summary not found in the database for session {inputs.session_id} when validating session summary with videos",
             non_retryable=True,
         )
+    # If the summary was already validated with videos, return
+    run_metadata = cast(dict[str, Any], summary_row.run_metadata)
+    if run_metadata and run_metadata.get("visual_confirmation"):
+        # Summary was already validated with videos, return
+        return None
     # Getting the user explicitly from the DB as we can't pass models between activities
-    user = await database_sync_to_async(User.objects.get)(id=inputs.user_id)
+    user = await User.objects.aget(id=inputs.user_id)
     if not user:
         raise ApplicationError(
             f"User not found in the database for user {inputs.user_id} when validating session summary with videos",
             non_retryable=True,
         )
     summary_row = cast(SingleSessionSummary, summary_row)
+    # TODO: Remove after tests
     with open(f"summary_{inputs.session_id}.json", "w") as f:
         json.dump(summary_row.summary, f, indent=4, sort_keys=True)
+    summary = SessionSummarySerializer(data=summary_row.summary)
+    summary.is_valid(raise_exception=True)
+    # Validate the session summary with videos
+    video_validator = SessionSummaryVideoValidator(
+        session_id=inputs.session_id,
+        summary=summary,
+        run_metadata=run_metadata,
+        team_id=inputs.team_id,
+        user=user,
+    )
+    updated_summary, updated_run_metadata = await video_validator.validate_session_summary_with_videos(
+        model_to_use=inputs.model_to_use
+    )
+    # Store the updated summary in the database
+    summary_row.summary = updated_summary.data
+    summary_row.run_metadata = asdict(updated_run_metadata)
+    await summary_row.asave(update_fields=["summary", "run_metadata"])
 
     # Pick events that happened before/after the exception, if they fit within the video duration
     # TODO: Decide if I need it
